+++ Vault/src/utopia/vault/database/columnlength/ColumnLengthRules.scala
@@ -1,6 +1,6 @@
 package utopia.vault.database.columnlength
 
-import utopia.flow.datastructure.immutable.DeepMap
+import utopia.flow.datastructure.immutable.{DeepMap, Model}
 import utopia.flow.generic.ValueConversions._
 import utopia.flow.parse.JsonParser
 import utopia.flow.util.StringExtensions._
@@ -92,26 +92,52 @@ object ColumnLengthRules
 	def loadFrom(path: Path)(implicit jsonParser: JsonParser, exc: ExecutionContext, connectionPool: ConnectionPool) =
 		jsonParser(path).map { json =>
 			val limits = json.getModel.attributes.flatMap { dbAtt =>
-				val dbName = dbAtt.name
-				dbAtt.value.getModel.attributes.flatMap { tableAtt =>
-					val tableName = tableAtt.name
-					tableAtt.value.getModel.attributes.flatMap { columnAtt =>
-						val propName = columnAtt.name
-						val value = columnAtt.value.getString.toLowerCase
-						val limit = value match {
-							case "throw" => Some(Throw)
-							case "crop" => Some(TryCrop)
-							case "expand" => Some(TryExpand.infinitely)
-							case _ =>
-								if (value.startsWith("expand") || value.startsWith("to"))
-									value.afterLast(" ").long.map { TryExpand upTo _ }
-								else
-									None
-						}
-						limit.map { Vector(dbName, tableName, propName) -> _ }
-					}
-				}
+				loadFromDbModel(dbAtt.name, dbAtt.value.getModel)
 			}
 			specifics ++= DeepMap(limits)
 		}
+	
+	/**
+	  * Loads column length rules from a .json file. The file should contain a json object with table names as
+	  * property names and objects as values. The nested objects should contain column properties.
+	  * The column properties should have values such as "throw", "crop",
+	  * "expand" or "expand to X" where X is the maximum allowed length.
+	  * @param path Path to the json file to read
+	  * @param databaseName Name of the database to which to apply these rules
+	  * @param jsonParser Json parser to use
+	  * @param exc Implicit execution context (used when handling connections opened during rule handling)
+	  * @param connectionPool Implicit connection pool (used for expanding column lengths when needed)
+	  * @return Success or failure, based on json parse result
+	  */
+	def loadFrom(path: Path, databaseName: String)
+	            (implicit jsonParser: JsonParser, exc: ExecutionContext, connectionPool: ConnectionPool) =
+	{
+		jsonParser(path).map { json =>
+			val limits = loadFromDbModel(databaseName, json.getModel)
+			specifics ++= DeepMap(limits)
+		}
+	}
+	
+	private def loadFromDbModel(dbName: String, model: Model)
+	                           (implicit exc: ExecutionContext, connectionPool: ConnectionPool) =
+	{
+		model.attributes.flatMap { tableAtt =>
+			val tableName = tableAtt.name
+			tableAtt.value.getModel.attributes.flatMap { columnAtt =>
+				val propName = columnAtt.name
+				val value = columnAtt.value.getString.toLowerCase
+				val limit = value match {
+					case "throw" => Some(Throw)
+					case "crop" => Some(TryCrop)
+					case "expand" => Some(TryExpand.infinitely)
+					case _ =>
+						if (value.startsWith("expand") || value.startsWith("to"))
+							value.afterLast(" ").long.map { TryExpand upTo _ }
+						else
+							None
+				}
+				limit.map { Vector(dbName, tableName, propName) -> _ }
+			}
+		}
+	}
 }