+++ Vault-Coder/Changes.md
@@ -5,12 +5,14 @@
 - `"description_link"` in the model structure document is now considered to refer to a class property and not a column
   - This may cause different results in cases where custom table names were used alongside the `"described"` property
 ### New Features
+- You can now write property length limit rules, which will be collected to a separate document (see README for details)
 - You can now customize column, table, property, etc. naming rules in the model declaration document 
   (see README for more details)
 - You can now customize id column name
 - You can now customize the name of the referenced column in the `ref` / `reference` property
-- You can now customize String and Int type maximum length / size by writing it into parentheses after the type name
+- You can now customize `String` and `Int` type maximum length / size by writing it into parentheses after the type name
   - See README for more details
+- You may now specify database name in the input document (see README for details)
 ### Other Changes
 - Class and property names within the documentation now follow standard writing style. 
   E.g. "word of life" instead of "wordOfLife"